image: remove image util package
